requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
plotly>=5.13.1
pdfplumber>=0.9.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.3
pyyaml>=6.0.1
bcrypt>=4.0.1 